app/requirements.txt
anyio==3.6.2
asyncpg==0.27.0
click==8.1.3
ecdsa==0.18.0
fastapi==0.95.2
greenlet==2.0.2
h11==0.14.0
idna==3.4
pyasn1==0.5.0
pydantic==1.10.7
pydub==0.25.1
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.14
starlette==0.27.0
typing_extensions==4.5.0
uvicorn==0.22.0
